requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
chardet==3.0.4
discord.py==1.0.0
idna==2.8
multidict==4.5.2
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
six==1.12.0
soupsieve==1.9
websockets==6.0
yarl==1.3.0
